Extracted simple GitHub client
